+++ examples/simple/priv/repo/migrations/20150115192936_create_weather.exs
@@ -7,6 +7,7 @@ defmodule Repo.CreatePosts do
       add :temp_lo, :integer
       add :temp_hi, :integer
       add :prcp,    :float
+      timestamps
     end
   end